rename kusto_ingest_client_exceptions to exceptions
