+++ azure-kusto-ingest/azure/kusto/ingest/__init__.py
@@ -1,6 +1,5 @@
 from .kusto_ingest_client import KustoIngestClient
 from .descriptors import BlobDescriptor, FileDescriptor
-from .kusto_ingest_client_exceptions import KustoDuplicateMappingError
 from .ingestion_properties import (
     DataFormat,
     ValidationPolicy,
