+++ azure-kusto-ingest/azure/kusto/ingest/exceptions.py
